unloadprotection to symbols
